Split useClientInvoicesPanel into domain hooks under 500 lines
Compose draft, invoice list, payments, and allocate/refund hooks from a thin
use-client-invoices-panel orchestrator. Shared banners, busy state, currency,
and cross-selection IDs live in dedicated modules with stable ref callbacks
for circular load dependencies. VM shape unchanged for finance subcomponents.
